backend/requirements.txt
Django==3.2.19
django_filter==22.1
djangorestframework-simplejwt>=5.2.2
djangorestframework==3.14.0
djoser==2.2.0
Pillow>=9.5.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.6
isort==5.12.0